refactor: split connector.rb into individual files
Split 241-line connector.rb into 11 one-class-per-file modules under
lib/xmi/sparx/connector/ with autoload. Pure file reorganization,
zero logic change.
